+++ src/main/java/ledger/user_interface/utils/AmountCreditStringConverter.java
@@ -5,14 +5,14 @@ package ledger.user_interface.utils;
  */
 public class AmountCreditStringConverter extends AAmountStringConverter {
 
-    public String toString(Integer amount) {
+    public String toString(Long amount) {
         if (amount == null) {
             return "";
         }
 
         String amountString = "";
         if (amount >= 0) {
-            Integer absoluteAmount = Math.abs(amount);
+            Long absoluteAmount = Math.abs(amount);
             String absoluteAmountInCents = String.valueOf(absoluteAmount);
 
             String dollars;
